Fix random loss of multicast, lasts 5-10 mins, by Ventus Networks
Once PIMD joins multicast groups it relies on a timer to periodically
send JOIN updates.  However, if JOIN update is received from a BSR
router (which is also a multicast router) then new timer + random value
is calculated.  If new timer’s value is greater than the value in the
packet (and in most cases it is) then the new value is used and next
JOIN prune update is delayed.  Delay typically starts at 80+ seconds.

The net result of this behavior is that JOIN message are actually never
get sent and after hold timer expiration the entire Linux PIM router is
removed from the group.  In practice this causes random outages that
lasts between 5 and 10 minutes.

Signed-off-by: Al Ivanov <alivanov@ventusgns.com>
Signed-off-by: Joachim Nilsson <troglobit@gmail.com>